removed gunship adds from LibBossIDs
The reason I added those adds there so that the Improvement module stores the Gunship battle. The problem with this approach is that those adds are considered bosses and they will be stored and we don't want that, we only want to store the Skybreaker or Ogrim's Hammer details, not those adds.