memory tweaks - Only what happens will be recorded
Using the "AddPlayerAttributes" and "AddSetAttributes" adds unnecessary data to tables that are either kept empty or cleaned using the "SetComplete" function. A better way to do things is to never record things if they don't happen and that's how it will be handled from now on on the following modules: CC, Dispels, Fails, FriendlyFire, Interrupts and Resurrects.
Example: if no dispels happened during an encounter, set.dispels and player.dispels tabels won't be created.