enhanced Enemies modules. fixes #31
There was previously a cache system that was used to collect data about enemies damage done or taken, this system prevent realtime update of numbers when clicking on an enemy bar.
This was changed, the cache system was removed and now real numbers are used instead and realtime update works fine (_test on solo, needs to be tested in raids_).