changed the way `Skada:IsBoss` works
When using Skada data from a different client language, the `Skada:IsBoss` function doesn't return any boss because the name is different.
To go around this, the name of the npc is passed as a second argument and used for comparison. If the name is different then we use, otherwise we use the name provided by LibBabble-Boss.