Update Auras.lua
We simply create a custom function that will round the player's active time in order to get kind of accurate auras time.
The reason we don't round up the player's active time during or after combat is because it's used to calculate DPS and HPS, so better keep the exact value of it and only round it down for Auras module.