end stopped segment after Smart Stop
The problem is not Smart Stop feature itself but how Skada "StopSegment" function works and how Skada saves segments.
After a smart stop, the segment is only saved if the player enters the next combat, this is problem because if the player reloads or relogs, the segment is gone. To prevent this, we make sure to end the segment once the player leaves combat, thus saving it.