fixed DBM/BigWigs Smart Stop issue
If you are doing OS3D for example, as soon as one of the Drakes dies DBM or BigWigs trigger the Smart Stop feature.
A check was added for it to not happen, the name of the killed unit must match the name of the Skada segment, which is takes default always the name of the boss.