correct Skada:ApplySettings behavior
The function was supposed to only apply settings to a window if provided, this prevents Skada from applying settings to everything causing useless amount of stuff. A check was added to make sure a window name is used in case of a provided window table.
In addition to this, a second param was added to the function that makes sure dropdown menus are closed after certain menu items were clicked.

Corrected some menus: what should close the menu closes it.