Goodbye LibWindow-1.1
LibWindow-1.1 is good, but not enough ... For some pixel-perfect people, positioning Skada windows is crucial and the library doesn't allow free-positioning, especially with the position optins added to windows. Thus, goodbye lib, we are going to use Skada's new function "SavePosition" and "RestorePosition" .. However, thanks for all the years you sticked around and served us.

+ removed few custom textures used by SpecializedLibBars and moved to using default game textures.
+ windows can be streched using the title button if locked..