optimize cache tables to reduce garbage
Cache tables are used to collect data that are deep in the SV table in order to build tables required for displays. Instead of reusing those tables, Skada used to create new ones each time some functions are called. This resulted in a lot of garbage ... what a shame!
This behavior was change and optimized and Skada now reuses those tables whenever required, causing tables to be created only if required, thus less garbage and less memory.