better boss fight detection
Skada used to always attempt to detect boss fights when in combat, the process is executed for every single CLEU event, which is irrelevant.
This behavior was changed so that the attempt to set boss fight is processed only once per target and only unique targets, so if you many targets of the same name, the process is executed only once.
As soon as a boss fight is detected, this process is completely eliminated and new "boss death" check starts, used to trigger "COMBAT_BOSS_DEFEATED" events.