move to LibDeflate for better compression
As of this commit, players with the recent update of the addon will use LibDeflate to export or transmit via addon comms.
As for receiving, a backward compatibility was  added so that profile importing or data receiving via comms still work, meaning players will the update as of this commit can receive comms from players with old versions, but not the other way around.