remove QueueUnit func and its relatives
Since this function and its relative were added only because Warmane bugs and only for the Death Knight's "Mark of Blood" healing not being attributed to the player but rather to the unit on which it was applied in order to properly attribute it. Thus, this context was moved to the healing module file instead, the issue should be handled by it.