LibCompat-1.0: move tContains from global namespace to lib
Overwriting `tContains` can cause unexpected behavior in Addons that depend on its old behavior and expect exactly `1` or `nil` as return and iterate the table as an array.
In addition, some custom servers change the interface code and use this function in protected code areas, and that overwrite will cause taint errors.